panelWidget: Do not override show() method
This widget defined a `show()` method, which would show the menu.
However, `show()` is already a method defined by some class higher up in
the hierarchy, which makes the *widget* in the status area visible.  By
defining a `show()` method here, the behavior of the `show()` method
changes, causing issues.

One such issue showed when combining this extension with the
status-area-horizontal-spacing extension, which calls `hide()` and
`show()` on each widget in the status area. With the overridden `show()`
method in the hamster extension, this would hide the widget and then
show the menu, breaking the extension because the widget would stay
hidden.

This commit renames `show()` to `show_menu()` to prevent to conflict.
It seems that the method was not actually used, so no references were
updated. For consistency, `toggle()` is also renamed to `toggle_menu()`,
with one reference updated.